address issue 2166 (was ok, actually)
I went back and forth on whether to halve the value in the parentheses
style, but then I read our docs and it was clearly stated that this
function follows the non-halved style.  So, I just added the ability to
set debug= in this function, cleaned up the docs a little, and am
closing this.

Frankly, I don't think this function is of much good, and I doubt that
many oce users are using this.

PS. I notice that the digits argument is ignored, at least for the
parentheses notation.  But I am not inclined to make an issue on that,
because (1) I wasted some time fiddling with ways to handle that and
there are a lot of edge cases so it is not straightforward and (2) I
don't care about that notation -- it's never used in papers/books that I
read, and there is disagreement in the two cited sources as to even what
is to be put there!