plotProfile() handles sigma0 with sigmalim (or xlim)
This was not working before, because I had (for some reason) shoved sigma[0:4] functionality into a category with other things, including u and v.  I suspect I did that as a quick fix. Frankly, this is a bit of a quick fix also because I want this kind of plot this morning.