format remaining files (misc.R manually...)
... I found that it just seemed to sit there for a long time when I
tried to format misc.R.  It may be that this has a LOT of blocks, and so
the formatter was having a hard time allocating memory.  Using 'top' did
not reveal what was going on.  In the end, I just used a vim macro to
fix up all the one-line blocks, and I changed a lot (not all) of the "{"
placements for functions.  So it is reasonably formatted, at least
enough so for my present purpose, and I will remember to be on the
lookout for formating problems there in future.  Maybe I should try
RStudio formatting...