reformat misc.R
The problem was in the example code for dataLabel().  When I removed the
example that contained @@ to show the slot accessor, it could then
format.  I learned this the hard way, by breaking this 6 kloc file down
in half, then breaking the bad half in half, etc. until I got to a short
(50 line?) chunk that would not format.  Some day I may try to boil that
down some more to make a bug report, but I actually don't know what vim
package is doing this work, so I don't know where to post an issue.