fix NumericVector size() format spec???
See https://github.com/dankelley/oce/issues/2172#issuecomment-1830700770
for the long and winding road I've been following on this.

I asked a question on the R-pkg-devel mailing list (which has had lots
of discussion of related matters in the past few days) and so I hope to
know the proper approach soon.  Lots of package maintainers must be
in the same boat, and surely some CRAN person will know about how
to do this to avoid problems on all or maybe just most R platforms.
And it's just a warning ... maybe CRAN will relax a bit until things are
clear and there is something that can work across platforms (hopefully
not an #ifdev though!).