give up on formatting NumericVector.size()
I am just not sure how to do it, and the advice I've had in the email
list doesn't seem to work.  If I ever want to return to this, I think
the name of this commit note ought to be enough, but below is a message
from the thread, but that R_PRIdLEN_T concatenation led to an error.
Maybe it's not defined in my toolchain.  But I am out of time to mess
with this. These are all occuring in rare circumstances of problems and
the point is that instead of saying things like "lengths of x and y must
match", these are stating those lengths. But the user can just check
that, if provided with the non-match message.  This is not worth
potentially having some CRAN machines issuing warnings that might
cause package rejection.  Over the next day or two, I'll continue local,
win_builder and R_CMD_CHECK tests, to try to find something that is
quiet in all three cases.  Then I'll submit oce.  Time is running out
and I have things to grade and exams coming up during the deadline
period.

::R_error("size %" R_PRIdXLEN_T ", a.size());

```
I think the best way to deal with these is to look at the lines and
arguments identified by any of the compilers on any of the test
platforms you have, and then based on reading the code and learning what
the types really are (or are guaranteed to be) decide on how to fix. The
presence and wording of the warnings may be compiler specific. In some
cases even platform specific, because some C types have different size
on different platforms, or map to different underlying types.

So, if the line above was identified by the compiler, check the
documentation for NumericVector::size(), which type does it guarantee to
return. The compiler warning may give a hint (what the type is in your
build), but indeed it wouldn't consult the documentation. If the type is
R_xlen_t (I didn't check if it is), then you could do e.g.

::R_error("size %lld", (long long)a.size());

or better

::R_error("size %" R_PRIdXLEN_T ", a.size());

as discussed elsewhere in this thread.

It is not a good idea to just use the type given in the warning, because
that may not be portable. In this case, your compiler warning (on Linux)
may have said "long int", which may have been how "ptrdiff_t" is defined
on the platform by the compiler, which is probably how R_xlen_t is
defined on your platform. But, it may be different elsewhere. On your
second platform, the compiler warning may have said "long long int" (on
Windows), which may have been how "ptrdiff_t" is defined there by the
compiler, and on that platform, "long int" is different from "long long
int", so the former didn't work.

If you use the Winbuilder service maintained by Uwe Ligges, you will
have the same setup as one of the platforms used for CRAN incoming checks.
```