I think large RDI files can now be handled...
The package passes the usual build-checks on my machine, which are a bit
more extensive than the officiala build-checks because I also include a
few more data files.  Having said that, the present code definitely
needs to be tested on large files, and I will ask for help on that from
the person who requested large-file support.