add fillGapMatrix()
This will be useful generally, and also it can be used in an updated
bin_average_2d() code.  My intention is to revert some of the changes to
bin_average_2d(), so that the core is done with cut() and split(), as it
has been for several months now (and as on CRAN).  All I need to do is
to use fillGapMatrix() to accomplish what I was trying to do over the
weekend, in revising C code from June 2023.

All of this is in support of issue 2199, which will not be solved for
several days, in my estimation.