remove unused C code for binning ...
This was added back in, temporarily, during work on issue 2199.  I
decided to remove it after I figured things out.  The only part of the
code we need in the present setup is now abstracted to another C++ file,
and the R wrapper is now generally available for any tasks requiring the
filling of NA gaps in matrices.