improve (and test) many [[ methods for seawater computations
This addresses issues #2207 and #2208.  See the checklist at the start
of the former, for an overview.  For more details -- since I didn't keep
the checklist up to date as I did the work -- look in the
tests/testthat/test_sw.R and tests/testthat/test_accessors.R test files.

I think all things are working now, and that I've trimmed the test
printing code I had inserted.  But, especially to guard against the
latter, I plan to use this branch in my daily work for a week or so,
before merging into the "develop" branch.