as.ctd(oce-class) gets lon and lat (issue #2237)
The problem was that the extraction of longitude and latitude was too
late in the code, so not all the special cases got it.  What are those
special cases?  Well, different things happen for the following types of
the first parameter:

* if is vector ...
* else if is list ...
* else if oce-class
    - if rsk-class ...
    - else if argo-class ...
    - else ...

I might have missed some, actually.  Anyway, lon and lat were captured
in all cases except the last, in the list above. My solution was to move
the capture to just after we detect oce-class, and before we start
looking for other things.

All the local tests pass, so I am pushing this to GH to get the 4 extra
remote tests that get done then.  If they are ok, I'll start the
external rhub tests (4 there, also, IIRC).  We just had a CRAN release,
so I'm not too worried about unexpected ramifications of the change,
since if such things turn up, we can just tell a user to revert from the
GH version to the CRAN version.