improve unit decoding (issue #2240)
We now handle quite a few things that write e.g. 'm-2' instead of
'/m^2'.  Also, the test suite checks every handled case.

But this is a steppingstone commit, because I want to rewrite this
complicated code (with it's hard-to-read regular expressions) as a
file-driven lookup table.  See
www.github.com/dankelley/oce-issues/2xxx/2240