igrf14 as downloaded from noaa.gov ...
It was downloaded with
    curl https://www.ngdc.noaa.gov/IAGA/vmod/igrf14.f > ~/git/oce/src/igrf14.f
and the next commit will be as edited to make it fit here. What we have
in the present commit will NOT compile.  The purpose of this commit, and
the next, is to explain exactly the changes that are required to port
updates to the magnetic field Fortran code for use in oce.