note project in R-CMD-check.yaml file
This makes it easier to compare two such files in an editor.  These
files are quite frustrating to have to deal with, because the various
remote scripts keep changing.  It seems that I need to do web searchers
on whether @v2 needs to be redone as @v3 (or whatever) about once a
year.  At about that frequency, we get errors on pushing to GH,