Change some size=2 to size=2L etc (maybe faster?)
This would save a conversion, so it might trim off a tiny bit of time.
Also, it's a bit clearer to the reader.  I see the L suffix a lot in
low-level R code and following a convention makes sense.

However, I only did this in one function (the ad2cp one).  My plan is
that if I'm editing a particular file, I'll take a look to see if there
are any things like this that could use polishing. Finding spots is easy
with e.g.

    grep -n "readBin.*size = [124]," file.R

and verifying the changes is easy with

    git diff --word-diff=color file.R