read.adp.ad2cp() docs cleanup byte-code refs
Also, remove old Nortek citations, which differed only in year, had no
URL provided, and were not referenced in the docs.  (Even the 2024
citation now has only a code block for the URL, because I found the URL
to be failing.  The only reason we mention the 2024 doc is because we
talk about the contradiction in the bytes relating to the number of
elements in altimeterRaw vectors.)