Move colour-related fcns from oce.R to colors.R
This is in preparation for adding CubeHelix.  Basically, I couldn't
guess where to put that code because I noticed that colour-related
functions were split across two files, oce.R and colors.R.  I've
wanted to consolidate this for quite a while, and now seems as good a
time as any. It builds, tests and checks okay.  This commit is so that
the next one, adding CubeHelix, will have a small diff.