Improve pwelch docs (a bit) BUT ...
... tomorrow I'm going to change the 'window' behaviour if 'nfft` and
'spec' are supplied.  (Yesterday I tried to fix bug 2299 by constructing
a Hamming window.  But that is NOT what is documented.  The docs say

"Note that no window will be applied to the data after subsampling, and
an error will be reported if window and spec are both given."

Given the age of the function (and the likelihood that this rule mimics
something in the Matlab code), I am inclined to obey that statement. In
the code, it just means that I'll use a flat window (as opposed to the
Hamming window I installed yesterday).  So it's an easy fix to the code.

More importantly, though, it would mean that the code will match the
documentation.

I am pretty sure that any user who knows enough to supply their own
spectral fit program is also fully able to do the windowing first.  This
is a power-user situation.

BUT ... I want to think about all of this.  I don't want to have to
modify the code in a few months if some user finds things are broken.