Improve pwelch (wrt issue #2299)
1. Disallow specifying 'window' if 'nfft' and 'spec' are provided.
2. Automatically use a Hamming window if 'nfft' and 'spec' are provided.
   (See github.com/dankelley/oce-issues/22xx/2299 for test code.)
3. Remove the local function that was in pwelch() for doing Hamming
   windowing.

Note that #3 benefits from extensions made to makeFilter(), so that it
can be used for both filtering and for windowing.  (These two tasks are
different in terms of normalization. In the filter case, consider
low-passing a nearly-constant time-series: we want the sum of
coefficients to equal 1, to avoid losing or gaining power.  However,
this is not what we want with a window.  A window usually is set up to
throw away data, e.g. in spectral computations, we do that to pull
endpoints down to zero or near-zero, to avoid getting the transform of a
tophat function, which is a sinc() function with all those bothersome
sidelobes. I won't explain further because folks who are likely to read
commit notes in detail are also likely to already be aware of the
difference between filtering and windowing.)