Revert in a GH action file...
I am reverting it because the advice given by the bot did not work.
Besides, the macOS test worked the last time I committed (Nov 21) so I
don't think the problem is with oce.  I think it's something on the
remote test machine.  Or with YACTFGWF (yet another change to effing
github work file).